feat(header): unify premium and normal member login button into shared component
- Removed duplicated components:
  - header/premium/premium-member-login-button.js
  - header/normal/member-login-button.js
- Added new shared `MemberLoginButton` component with `variant` prop to handle both cases
