+++ packages/mirror-media-next/components/header/normal/header.js
@@ -13,7 +13,7 @@ import { DEFAULT_NORMAL_SECTIONS_DATA } from '../../../constants/header'
 import SubscribeMagazineButton from './subscribe-papermag-mb-btn'
 import EntranceButton from './entrance-button'
 import Logo from './logo'
-import MemberLoginButton from './member-login-button'
+import MemberLoginButton from '../shared/member-Login-Button'
 import MobileSidebar from './mobile-sidebar'
 import NavSections from './nav-sections'
 import NavTopics from './nav-topics'
