+++ packages/mirror-media-next/components/header/premium/premium-header.js
@@ -8,8 +8,8 @@ import SearchBarInput from '../shared/search-bar-input'
 import Logo from '../normal/logo'
 import PremiumMobileSidebar from './premium-mobile-sidebar'
 import PremiumNavSections from './premium-nav-sections'
-import PremiumMemberLoginButton from './premium-member-login-button'
 import useSearch from '../../../hooks/use-search'
+import MemberLoginButton from '../shared/member-Login-Button'
 
 /**
  * @typedef {import('./premium-mobile-sidebar').H2AndH3Block} H2AndH3Block
@@ -126,7 +126,7 @@ const SearchInputWrapper = styled.div`
   }
 `
 
-const DesktopPremiumMemberLoginButton = styled(PremiumMemberLoginButton)`
+const DesktopPremiumMemberLoginButton = styled(MemberLoginButton)`
   display: none;
   ${({ theme }) => theme.breakpoint.xl} {
     display: inline-flex;
@@ -145,9 +145,8 @@ const DesktopPremiumMemberLoginButton = styled(PremiumMemberLoginButton)`
     }
   }
 `
-const MobilePremiumMemberLoginButton = styled(PremiumMemberLoginButton)`
+const MobilePremiumMemberLoginButton = styled(MemberLoginButton)`
   display: block;
-  margin-left: 15px;
   ${({ theme }) => theme.breakpoint.xl} {
     display: none;
   }
@@ -238,7 +237,7 @@ export default function PremiumHeader({
               alt="search-button"
             />
           </SearchButtonMobile>
-          <MobilePremiumMemberLoginButton />
+          <MobilePremiumMemberLoginButton variant="premium" />
           <PremiumMobileSidebar
             sections={sections}
             shouldShowSubtitleNavigator={shouldShowSubtitleNavigator}
@@ -270,7 +269,7 @@ export default function PremiumHeader({
           />
         </SearchInputWrapper>
         <PremiumNavSections sections={sections}>
-          <DesktopPremiumMemberLoginButton />
+          <DesktopPremiumMemberLoginButton variant="premium" />
         </PremiumNavSections>
       </HeaderBottom>
     </HeaderWrapper>
